+++ scripts/macros.js
@@ -302,6 +302,7 @@ import {wardingBond} from './macros/spells/wardingBond.js';
 import {wildShape} from './macros/classFeatures/druid/wildshape.js';
 import {wildhunt} from './macros/raceFeatures/shifter/wildhunt.js';
 import {witherAndBloom} from './macros/spells/witherAndBloom.js';
+import {wrapsOffDyamak} from './macros/items/wrapsOffDyamak.js';
 import {wrathOfTheStorm} from './macros/classFeatures/cleric/tempestDomain/wrathOfTheStorm.js';
 import {zealousPresence} from './macros/classFeatures/barbarian/zealot/zealousPresence.js';
 import {zombie} from './macros/monsterFeatures/zombie/zombie.js';
@@ -705,6 +706,7 @@ export let macros = {
     'wildShape': wildShape,
     'wildhunt': wildhunt,
     'witherAndBloom': witherAndBloom,
+    'wrapsOffDyamak': wrapsOffDyamak,
     'wrathOfTheStorm': wrathOfTheStorm,
     'zealousPresence': zealousPresence,
     'zoneOfTruth': zoneOfTruth