+++ scripts/macros/classFeatures/cleric/tempestDomain/thunderboltStrike.js
@@ -1,10 +1,10 @@
+import {constants} from '../../../../constants.js';
 import {chris} from '../../../../helperFunctions.js';
 import {queue} from '../../../../utility/queue.js';
 export async function thunderboltStrike({speaker, actor, token, character, item, args, scope, workflow}) {
-    if (workflow.hitTargets.size === 0 || !workflow.damageRoll) return;
-    let targetToken = workflow.targets.first();
-    let targetActor = targetToken.actor;
-    if (!(targetActor.system.traits.size === 'lg' || targetActor.system.traits.size === 'med' || targetActor.system.traits.size === 'sm' || targetActor.system.traits.size === 'tiny')) return;
+    if (!workflow.hitTargets.size || !workflow.damageRoll) return;
+    let feature = chris.getItem(workflow.actor, 'Thunderbolt Strike');
+    if (!feature) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'thunderboltStrike', 475);
     if (!queueSetup) return;
     let damageTypes = chris.getRollDamageTypes(workflow.damageRoll);
@@ -12,41 +12,21 @@ export async function thunderboltStrike({speaker, actor, token, character, item,
         queue.remove(workflow.item.uuid);
         return;
     }
-    let selection = await chris.dialog('Thunderbolt Strike: Push target?', [['Yes', 10], ['No', false]]);
-    if (!selection) {
+    let targets = Array.from(workflow.hitTargets).filter(i => chris.getSize(i.actor) <= 3);
+    if (!targets.length) {
         queue.remove(workflow.item.uuid);
         return;
     }
-    let knockBackFactor;
-    let ray;
-    let newCenter;
-    let hitsWall = true;
-    while (hitsWall) {
-        knockBackFactor = selection / canvas.dimensions.distance;
-        ray = new Ray(workflow.token.center, targetToken.center);
-        newCenter = ray.project(1 + ((canvas.dimensions.size * knockBackFactor) / ray.distance));
-        hitsWall = targetToken.checkCollision(newCenter, {origin: ray.A, type: "move", mode: "any"});
-        if (hitsWall) {
-            selection -= 5;
-            if (selection === 0) {
-                ui.notifications.info('Target is unable to be moved!');
-                queue.remove(workflow.item.uuid);
-                return;
-            }
-        }
+    let selection = await chris.selectTarget(feature.name + ': Push targets?', constants.yesNoButton, targets, true, 'multiple');
+    if (!selection.buttons) {
+        queue.remove(workflow.item.uuid);
+        return;
     }
-    newCenter = canvas.grid.getSnappedPosition(newCenter.x - targetToken.w / 2, newCenter.y - targetToken.h / 2, 1);
-    let targetUpdate = {
-        'token': {
-            'x': newCenter.x,
-            'y': newCenter.y
-        }
-    };
-    let options = {
-        'permanent': true,
-        'name': workflow.item.name,
-        'description': workflow.item.name
-    };
-    await warpgate.mutate(targetToken.document, targetUpdate, {}, options);
+    let pushTargets = selection.inputs.map(i => fromUuidSync(i).object);
+    if (!pushTargets.length) {
+        queue.remove(workflow.item.uuid);
+        return;
+    }
+    for (let i of pushTargets) chrisPremades.helpers.pushToken(workflow.token, i, 10);
     queue.remove(workflow.item.uuid);
 }